♻️(commons-domain) rename audit domain and django_apps into commons
