+++ src/web/infrastructure/django_apps/commons/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class AuditConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "infrastructure.django_apps.audit"
+    name = "infrastructure.django_apps.commons"
