+++ src/web/tests/utils/shared_fixtures.py
@@ -38,13 +38,13 @@ from application.ingestion.usecases.load_documents import LoadDocumentsUsecase
 from application.ingestion.usecases.upsert_offers import UpsertOffersUseCase
 from application.ingestion.usecases.vectorize_documents import VectorizeDocumentsUsecase
 from config.app_config import AppConfig
-from domain.audit.services.audit_log_writer import AuditLogWriter
 from domain.candidate.repositories.candidature_repository_interface import (
     ICandidatureRepository,
 )
 from domain.candidate.repositories.cv_metadata_repository_interface import (
     ICVMetadataRepository,
 )
+from domain.commons.services.audit_log_writer import AuditLogWriter
 from domain.identite.repositories.agent_repository_interface import IAgentRepository
 from domain.identite.repositories.utilisateur_repository_interface import (
     IUtilisateurRepository,
